Adding React Component
