+++ index.html
@@ -4,74 +4,12 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./styles/index.css" />
     <title>Document</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400&family=Readex+Pro&display=swap" rel="stylesheet"
-    />
+    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400&family=Readex+Pro&display=swap" rel="stylesheet" />
   </head>
   <body>
-    <div class="feedback slider">
-      <div class="slide">
-        <a class="prev" onclick="plusSlides(-1)">&#10094;</a>
-        <div class="content">
-          <h1>Leila Domniuc</h1>
-          <h2>Head of marketing,Facebook</h2>
-          <p class="rating">⭐⭐⭐⭐</p>
-          <p class="review">
-            My project was a simple task,but the persistence of ladidapp web
-            development team turend an very awesome and grear project make me
-            very happy Thank you so much.
-          </p>
-        </div>
-        <div class="forImage">
-          <img src="images/person.png" />
-        </div>
-
-        <a class="next" onclick="plusSlides(1)">&#10095;</a>
-      </div>
-      <div class="slide">
-        <a class="prev" onclick="plusSlides(-1)">&#10094;</a>
-
-        <div class="content">
-          <h1>Markin nesus</h1>
-          <h2>Digital Inc</h2>
-          <p class="rating">⭐⭐⭐⭐</p>
-          <p class="review">
-            My project was a simple task,but the persistence of ladidapp web
-            development team turend an very awesome and grear project make me
-            very happy Thank you so much.
-          </p>
-        </div>
-
-        <div class="forImage">
-          <img src="images/person.png" />
-        </div>
-
-        <a class="next" onclick="plusSlides(1)">&#10095;</a>
-      </div>
-
-      <div class="slide">
-        <a class="prev" onclick="plusSlides(-1)">&#10094;</a>
-        <div class="content">
-          <h1>John Peter</h1>
-          <h2>Google Inc</h2>
-          <p class="rating">⭐⭐⭐⭐</p>
-          <p class="review">
-            My project was a simple task, but the persistence of ladidapp web
-            development team turend an very awesome and grear project make me
-            very happy Thank you so much.
-          </p>
-        </div>
-
-        <div class="forImage">
-          <img src="images/person.png" />
-        </div>
-
-        <a class="next" onclick="plusSlides(1)">&#10095;</a>
-      </div>
-    </div>
-    <script src="./scripts/index.js"></script>
+    <div id="root"></div>
   </body>
 </html>
